+++ example_gradle_8/android/app/src/main/kotlin/com/flutter/instoreappversionchecker/example/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.example_gradle_8
+package com.flutter.instoreappversionchecker.example
 
 import io.flutter.embedding.android.FlutterActivity